src/requirements.txt
tink>=1.7.0
cryptography>=41.0.0
requests>=2.31.0
langchain
langchain-community
langchain-ollama
langchain-chroma
streamlit
ollama
httpx
chromadb>=0.5.23
fastapi
uvicorn
python-multipart
sentence-transformers>=2.2.0
torch>=2.0.0
pillow>=10.0.0
